fix(showcase/langgraph-typescript): align open-gen-ui to LGP gold standard
Copy all LGP files exactly: page.tsx, chat.tsx, design-skill.ts,
suggestions.ts.
